ansaf backend/requirements.txt
django==5.2.6
djangorestframework==3.15.2
django-filter==24.3
django-taggit==6.0.0
Pillow==10.4.0
psycopg[binary]==3.2.1
django-storages==1.14.4
python-dotenv==1.0.1
drf-spectacular==0.27.2
bleach==6.1.0
channels==4.1.0
channels-redis==4.2.0
websockets==15.0.1
daphne==4.1.2
django-cors-headers==4.4.0